+++ changelog.md
@@ -9,3 +9,18 @@ Changelog format: [Markdown](https://github.com/adam-p/markdown-here/wiki/Markdo
 Release Date: Sep 30, 2017
 
 - Initialized the Android project
+
+## Version 0.0.2
+
+Release Date: Oct 11, 2017
+
+- Adding Splash Activity (Application loading screen)
+- Adding Homepage Activity
+- Adding logo to the app
+
+## Version 0.0.3
+
+Release Date: Oct 19, 2017
+
+- Adding History Activity
+